@@ -9,30 +9,65 @@ plugins {
     id("com.getkeepsafe.dexcount")
 }
 
+apply(plugin = "com.google.gms.google-services")
+
 android {
-    compileSdkVersion(27)
-    buildToolsVersion = "27.0.3"
+    compileSdkVersion(28)
+    buildToolsVersion = "28.0.3"
 
     defaultConfig {
         applicationId = "hr.bpervan.novaeva.main"
         minSdkVersion(16)
-        targetSdkVersion(27)
-        versionCode = 160020905
-        versionName = "2.9.5"
+        targetSdkVersion(28)
+        versionCode = 160030100
+        versionName = "3.1.0"
     }
 
     buildTypes {
         getByName("debug") {
             isShrinkResources = false
             isMinifyEnabled = true
-            isUseProguard = false
             proguardFiles(getDefaultProguardFile("proguard-android.txt"), "proguard-rules.pro")
         }
         getByName("release") {
             isShrinkResources = true
             isMinifyEnabled = true
-            isUseProguard = true
             proguardFiles(getDefaultProguardFile("proguard-android.txt"), "proguard-rules.pro")
+        }
+    }
+
+    flavorDimensions("server")
+
+    productFlavors {
+        create("development") {
+            dimension = "server"
+
+            applicationIdSuffix = ".dev"
+            versionNameSuffix = "-dev"
+
+            resValue("string", "app_name", "Nova Deva")
+
+            buildConfigField("String", "V2_SERVER_URI", "\"http://novaeva.com\"")
+            buildConfigField("String", "V3_SERVER_URI", "\"https://api.hr.novaeva.xyz\"")
+            buildConfigField("String", "V3_SERVER_AUTH", "\"Basic cmFESW8wMDE6OUtoSjk5MlZSQ0d1bU5hRzlTeWI=\"")
+            buildConfigField("String", "DB_NAME", "\"NovaEvaDb_dev.realm\"")
+
+            buildConfigField("long", "V2_LIVE_ID", "202")
+            buildConfigField("long", "V3_LIVE_ID", "2724")
+        }
+
+        create("production") {
+            dimension = "server"
+
+            resValue("string", "app_name", "Nova Eva")
+
+            buildConfigField("String", "V2_SERVER_URI", "\"http://novaeva.com\"")
+            buildConfigField("String", "V3_SERVER_URI", "\"https://api.hr.novaeva.com\"")
+            buildConfigField("String", "V3_SERVER_AUTH", "\"Basic QW5kUm9pRDAwMTpkM2JXYXdnRFpqcXFWV2dESlJLag==\"")
+            buildConfigField("String", "DB_NAME", "\"NovaEvaDb_prod.realm\"")
+
+            buildConfigField("long", "V2_LIVE_ID", "202")
+            buildConfigField("long", "V3_LIVE_ID", "2724")
         }
     }
 
