@@ -1,7 +1,7 @@
 package hr.bpervan.novaeva.adapters
 
-import android.support.v4.content.ContextCompat
-import android.support.v7.widget.RecyclerView
+import androidx.core.content.ContextCompat
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -9,7 +9,7 @@ import android.widget.TextView
 import hr.bpervan.novaeva.EventPipelines
 import hr.bpervan.novaeva.adapters.RadioStationsAdapter.RadioStationViewHolder
 import hr.bpervan.novaeva.main.R
-import hr.bpervan.novaeva.model.EvaContentMetadata
+import hr.bpervan.novaeva.model.EvaContent
 import hr.bpervan.novaeva.util.EvaTouchFeedback
 import hr.bpervan.novaeva.views.PlayPauseView
 import kotlinx.android.synthetic.main.recycler_item_radio_station.view.*
@@ -17,8 +17,8 @@ import kotlinx.android.synthetic.main.recycler_item_radio_station.view.*
 /**
  *
  */
-class RadioStationsAdapter(private val radioStations: List<EvaContentMetadata>)
-    : RecyclerView.Adapter<RadioStationViewHolder>() {
+class RadioStationsAdapter(private val radioStations: List<EvaContent>)
+    : androidx.recyclerview.widget.RecyclerView.Adapter<RadioStationViewHolder>() {
 
     override fun getItemCount() = radioStations.size
 
@@ -26,7 +26,7 @@ class RadioStationsAdapter(private val radioStations: List<EvaContentMetadata>)
 
     private var touchFeedbackColor: Int = 0
 
-    override fun onAttachedToRecyclerView(recyclerView: RecyclerView) {
+    override fun onAttachedToRecyclerView(recyclerView: androidx.recyclerview.widget.RecyclerView) {
         super.onAttachedToRecyclerView(recyclerView)
 
         touchFeedbackColor = ContextCompat.getColor(recyclerView.context, R.color.TranslucentGray)
@@ -42,18 +42,18 @@ class RadioStationsAdapter(private val radioStations: List<EvaContentMetadata>)
         holder.bindTo(radioStations[position])
     }
 
-    inner class RadioStationViewHolder(val view: View) : RecyclerView.ViewHolder(view) {
+    inner class RadioStationViewHolder(val view: View) : androidx.recyclerview.widget.RecyclerView.ViewHolder(view) {
         private val stationName: TextView = view.stationName
         private val imgPlayPause: PlayPauseView = view.imgPlayPause
 
-        fun bindTo(radioStation: EvaContentMetadata) {
+        fun bindTo(radioStation: EvaContent) {
             stationName.text = radioStation.title
             view.setOnTouchListener(EvaTouchFeedback(view, touchFeedbackColor))
             view.setOnClickListener {
                 EventPipelines.chooseRadioStation.onNext(radioStation)
             }
 
-            imgPlayPause.isPlaying = radioStation.contentId == radioStationPlaying
+            imgPlayPause.isPlaying = radioStation.id == radioStationPlaying
         }
     }
 }
