@@ -2,18 +2,25 @@ package hr.bpervan.novaeva.player
 
 import android.content.Context
 import android.content.Intent
-import android.support.v4.content.ContextCompat
+import android.os.Handler
+import androidx.core.content.ContextCompat
+import androidx.core.net.toUri
 import com.google.android.exoplayer2.ExoPlayer
 import com.google.android.exoplayer2.ExoPlayerFactory
 import com.google.android.exoplayer2.Player
+import com.google.android.exoplayer2.source.ExtractorMediaSource
 import com.google.android.exoplayer2.source.MediaSource
 import com.google.android.exoplayer2.trackselection.AdaptiveTrackSelection
 import com.google.android.exoplayer2.trackselection.DefaultTrackSelector
 import com.google.android.exoplayer2.ui.PlayerView
 import com.google.android.exoplayer2.upstream.DefaultBandwidthMeter
+import com.google.android.exoplayer2.upstream.DefaultHttpDataSourceFactory
+import com.google.android.exoplayer2.util.Util
+import hr.bpervan.novaeva.EventPipelines
+import hr.bpervan.novaeva.EventPipelines.playbackStartStopPause
 import hr.bpervan.novaeva.NovaEvaApp
+import hr.bpervan.novaeva.main.R
 import hr.bpervan.novaeva.services.AudioPlayerService
-import io.reactivex.subjects.BehaviorSubject
 
 /**
  *
@@ -24,7 +31,8 @@ class EvaPlayer(context: Context) {
                          val playbackInfo: PlaybackInfo?)
 
     class PlaybackInfo(val id: String,
-                       val title: String? = null)
+                       val title: String?,
+                       val isRadio: Boolean)
 
     private val playerAlpha: ExoPlayer = createDefaultExoPlayer(context)
     private val playerBeta: ExoPlayer = createDefaultExoPlayer(context)
@@ -32,8 +40,6 @@ class EvaPlayer(context: Context) {
     private val playerPlaybackInfoMap = mutableMapOf<ExoPlayer, PlaybackInfo?>()
 
     init {
-        ContextCompat.startForegroundService(context, Intent(context, AudioPlayerService::class.java))
-
         playerAlpha.addListener(StartServiceIfNeeded(playerAlpha, playerBeta))
         playerBeta.addListener(StartServiceIfNeeded(playerBeta, playerAlpha))
     }
@@ -42,55 +48,93 @@ class EvaPlayer(context: Context) {
                                      private val otherPlayer: ExoPlayer) : Player.DefaultEventListener() {
 
         override fun onPlayerStateChanged(playWhenReady: Boolean, playbackState: Int) {
-            if ((playbackState == Player.STATE_READY)) {
-                if (playWhenReady) {
+            if ((thisPlayer.playbackState == Player.STATE_READY)) {
+                if (thisPlayer.playWhenReady) {
 
                     otherPlayer.playWhenReady = false
                     otherPlayer.stop(true)
-                }
-                val ctx = NovaEvaApp.instance ?: return
 
-                ContextCompat.startForegroundService(ctx, Intent(ctx, AudioPlayerService::class.java))
-            } else if (playbackState == Player.STATE_IDLE) {
+                    val ctx = NovaEvaApp.instance ?: return
+                    ContextCompat.startForegroundService(ctx, Intent(ctx, AudioPlayerService::class.java))
+                }
+                if (otherPlayer.playbackState == Player.STATE_IDLE
+                        || otherPlayer.playbackState == Player.STATE_ENDED
+                        || otherPlayer.playWhenReady.not()) {
+                    emitPlaybackStartStopPause()
+                }
+            } else if (playbackState == Player.STATE_IDLE
+                    || playbackState == Player.STATE_ENDED) {
+
                 playerPlaybackInfoMap[thisPlayer] = null
+
+                if (otherPlayer.playbackState == Player.STATE_IDLE
+                        || otherPlayer.playbackState == Player.STATE_ENDED
+                        || otherPlayer.playWhenReady.not()) {
+                    emitPlaybackStartStopPause()
+                }
             }
-            emitPlaybackChange()
         }
 
-        private fun emitPlaybackChange() {
-            playbackChangeSubject.onNext(PlaybackChange(thisPlayer, playerPlaybackInfoMap[thisPlayer]))
+        private fun emitPlaybackStartStopPause() {
+            playbackStartStopPause.onNext(PlaybackChange(thisPlayer, playerPlaybackInfoMap[thisPlayer]))
         }
     }
 
-    val playbackChangeSubject = BehaviorSubject.create<PlaybackChange>()
+    private val handler = Handler()
 
-    fun prepareIfNeeded(playbackInfo: PlaybackInfo, doAutoPlay: Boolean = false, mediaSourceProvider: () -> MediaSource) {
+    fun prepareAudioStream(audioUri: String,
+                           contentId: String,
+                           contentTitle: String,
+                           isRadio: Boolean,
+                           doAutoPlay: Boolean,
+                           auth: String?) {
+
+        val context = NovaEvaApp.instance ?: return
+
+        prepareIfNeeded(EvaPlayer.PlaybackInfo(contentId, contentTitle, isRadio), doAutoPlay) {
+
+            val dataSourceFactory = DefaultHttpDataSourceFactory(
+                    Util.getUserAgent(context, context.resources.getString(R.string.app_name)),
+                    DefaultBandwidthMeter())
+
+            if (auth != null) {
+                dataSourceFactory.defaultRequestProperties.set("Authorization", auth)
+            }
+
+            ExtractorMediaSource.Factory(dataSourceFactory).createMediaSource(audioUri.toUri(), handler, null)
+        }
+    }
+
+    private fun Player.isPlaying(): Boolean {
+        return (playbackState == Player.STATE_READY
+                || playbackState == Player.STATE_BUFFERING) && playWhenReady
+    }
+
+    private fun Player.isStopped(): Boolean {
+        return playbackState == Player.STATE_IDLE
+                || playbackState == Player.STATE_ENDED
+    }
+
+    private fun prepareIfNeeded(playbackInfo: PlaybackInfo, doAutoPlay: Boolean = false, mediaSourceProvider: () -> MediaSource) {
 
         val currentPlayer = if (playerBeta.isPlaying()) playerBeta else playerAlpha
         val otherPlayer = if (currentPlayer == playerAlpha) playerBeta else playerAlpha
 
-        when {
-            playbackInfo.id == playerPlaybackInfoMap[currentPlayer]?.id -> {
-                if (doAutoPlay) {
-                    currentPlayer.playWhenReady = true
-                }
+        if (playbackInfo.id == playerPlaybackInfoMap[currentPlayer]?.id) {
+            if (doAutoPlay) {
+                currentPlayer.playWhenReady = true
             }
-            currentPlayer.isStopped() || doAutoPlay -> {
-                currentPlayer.apply {
-                    playWhenReady = false
-                    prepare(mediaSourceProvider())
-                    playWhenReady = doAutoPlay
-                }
-                playerPlaybackInfoMap[currentPlayer] = playbackInfo
+            return
+        }
 
-            }
-            else -> {
-                otherPlayer.apply {
-                    playWhenReady = false
-                    prepare(mediaSourceProvider())
-                }
-                playerPlaybackInfoMap[otherPlayer] = playbackInfo
-            }
+        val playerToUse = if (currentPlayer.isStopped()) currentPlayer else otherPlayer
+
+        playerToUse.apply {
+            playWhenReady = false
+            stop(true)
+            prepare(mediaSourceProvider())
+            playerPlaybackInfoMap[this] = playbackInfo
+            playWhenReady = doAutoPlay
         }
     }
 
@@ -106,13 +150,18 @@ class EvaPlayer(context: Context) {
         playerBeta.stop()
     }
 
+    fun isStopped(): Boolean {
+        return playerAlpha.playbackState == Player.STATE_IDLE
+                && playerBeta.playbackState == Player.STATE_IDLE
+    }
+
     fun pause() {
         playerAlpha.playWhenReady = false
         playerBeta.playWhenReady = false
     }
 
     fun currentPlaybackInfo(): PlaybackInfo? {
-        return NovaEvaApp.evaPlayer.playbackChangeSubject.value?.playbackInfo
+        return EventPipelines.playbackStartStopPause.value?.playbackInfo
     }
 
     fun supplyPlayerToView(playerView: PlayerView, playbackId: String) {
