@@ -1,21 +1,26 @@
 package hr.bpervan.novaeva.fragments
 
 import android.os.Bundle
-import android.support.design.widget.Snackbar
+import android.util.Log
 import android.view.ContextThemeWrapper
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.core.os.bundleOf
-import com.google.android.gms.analytics.HitBuilders
+import com.google.firebase.analytics.FirebaseAnalytics
 import hr.bpervan.novaeva.EventPipelines
 import hr.bpervan.novaeva.NovaEvaApp
-import hr.bpervan.novaeva.util.SCROLL_PERCENT_KEY
 import hr.bpervan.novaeva.main.R
-import hr.bpervan.novaeva.services.novaEvaService
-import hr.bpervan.novaeva.util.networkRequest
+import hr.bpervan.novaeva.rest.NovaEvaService
+import hr.bpervan.novaeva.util.BREVIARY_IMAGE_KEY
+import hr.bpervan.novaeva.util.dataErrorSnackbar
 import hr.bpervan.novaeva.util.plusAssign
-import hr.bpervan.novaeva.views.*
+import hr.bpervan.novaeva.views.applyConfiguredFontSize
+import hr.bpervan.novaeva.views.applyEvaConfiguration
+import hr.bpervan.novaeva.views.loadHtmlText
+import io.reactivex.android.schedulers.AndroidSchedulers
+import io.reactivex.rxkotlin.subscribeBy
+import io.reactivex.schedulers.Schedulers
 import kotlinx.android.synthetic.main.collapsing_content_header.view.*
 import kotlinx.android.synthetic.main.fragment_simple_content.*
 
@@ -50,7 +55,7 @@ class BreviaryContentFragment : EvaBaseFragment() {
 
         val inState = savedInstanceState ?: arguments!!
         breviaryId = inState.getInt(BREVIARY_ID_KEY, 4)
-        coverImageUrl = prefs.getString("hr.bpervan.novaeva.brevijarheaderimage", null)
+        coverImageUrl = prefs.getString(BREVIARY_IMAGE_KEY, null)
         breviaryName = "Brevijar - " + when (breviaryId) {
             1 -> "Jučer, Jutarnja"
             2 -> "Jučer, Večernja"
@@ -62,14 +67,6 @@ class BreviaryContentFragment : EvaBaseFragment() {
             8 -> "Sutra, Večernja"
             else -> "Sutra, Povečerje"
         }
-
-        savedInstanceState ?: NovaEvaApp.defaultTracker
-                .send(HitBuilders.EventBuilder()
-                        .setCategory("Brevijar")
-                        .setAction("OtvorenBrevijar")
-                        .setLabel(breviaryName)
-                        .setValue(breviaryId.toLong())
-                        .build())
     }
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
@@ -85,13 +82,6 @@ class BreviaryContentFragment : EvaBaseFragment() {
         EventPipelines.changeStatusbarColor.onNext(R.color.Transparent)
         EventPipelines.changeFragmentBackgroundResource.onNext(R.color.Transparent)
 
-        val savedScrollPercent = savedInstanceState?.getFloat(SCROLL_PERCENT_KEY, 0f) ?: 0f
-        if (savedScrollPercent > 0) {
-            webView.afterLoadAndLayoutComplete {
-                simpleContentScrollView.scrollY = calcScrollYAbsolute(savedScrollPercent, webView.height)
-            }
-        }
-
         if (savedInstanceState != null && savedBreviaryText != null) {
             breviaryText = savedBreviaryText
             showBreviary()
@@ -99,10 +89,10 @@ class BreviaryContentFragment : EvaBaseFragment() {
             fetchBreviary()
         }
 
-        initUI()
-    }
+        disposables += EventPipelines.resizeText.subscribe {
+            webView?.applyConfiguredFontSize(prefs)
+        }
 
-    private fun initUI() {
         webView.applyEvaConfiguration(prefs)
 
         evaCollapsingBar.collapsingToolbar.title = breviaryName
@@ -114,23 +104,32 @@ class BreviaryContentFragment : EvaBaseFragment() {
         }
     }
 
+    override fun onResume() {
+        super.onResume()
+
+        FirebaseAnalytics.getInstance(requireContext())
+                .setCurrentScreen(requireActivity(), breviaryName.take(36), "BreviaryContent")
+    }
+
     override fun onSaveInstanceState(outState: Bundle) {
         outState.putInt(BREVIARY_ID_KEY, breviaryId)
-        outState.putFloat(SCROLL_PERCENT_KEY, simpleContentScrollView.calcScrollYPercent(webView.height))
         savedBreviaryText = breviaryText
         super.onSaveInstanceState(outState)
     }
 
     private fun fetchBreviary() {
 
-        baseDisposables += novaEvaService
+        disposables += NovaEvaService.v2
                 .getBreviary(breviaryId.toString())
-                .networkRequest({ breviary ->
-                    view ?: return@networkRequest
+                .subscribeOn(Schedulers.io())
+                .observeOn(AndroidSchedulers.mainThread())
+                .subscribeBy(onSuccess = { breviary ->
+                    view ?: return@subscribeBy
                     breviaryText = breviary.text ?: ""
                     showBreviary()
                 }, onError = {
-                    view?.snackbar(R.string.error_fetching_data, Snackbar.LENGTH_LONG)
+                    Log.e("breviaryError", it.message, it)
+                    view?.dataErrorSnackbar()
                 })
     }
 
