@@ -2,13 +2,15 @@ package hr.bpervan.novaeva.fragments
 
 import android.content.SharedPreferences
 import android.os.Bundle
-import android.support.v4.app.Fragment
 import android.view.View
+import com.google.android.exoplayer2.Player
 import com.nostra13.universalimageloader.core.ImageLoader
 import hr.bpervan.novaeva.EventPipelines
 import hr.bpervan.novaeva.NovaEvaApp
 import hr.bpervan.novaeva.util.minusAssign
 import hr.bpervan.novaeva.util.plusAssign
+import hr.bpervan.novaeva.views.EvaRadioBtn
+import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.disposables.Disposable
 import kotlinx.android.synthetic.main.fragment_prayers.*
@@ -16,12 +18,12 @@ import kotlinx.android.synthetic.main.fragment_prayers.*
 /**
  * Created by vpriscan on 26.11.17..
  */
-abstract class EvaBaseFragment : Fragment() {
+abstract class EvaBaseFragment : androidx.fragment.app.Fragment() {
 
     /**
      * Fragment's lifecycle bound disposables
      */
-    protected val baseDisposables = CompositeDisposable()
+    protected val disposables = CompositeDisposable()
 
     protected val prefs: SharedPreferences
         get() = NovaEvaApp.prefs
@@ -29,9 +31,14 @@ abstract class EvaBaseFragment : Fragment() {
     protected val imageLoader: ImageLoader
         get() = NovaEvaApp.imageLoader
 
-    interface EvaFragmentFactory<out T : Fragment, in K> {
+    interface EvaFragmentFactory<out T : androidx.fragment.app.Fragment, in K> {
+
+        companion object {
+            const val INITIALIZER = "initializer"
+        }
+
         val tag: String
-            get() = javaClass.canonicalName
+            get() = javaClass.canonicalName!!
 
         fun newInstance(initializer: K): T
     }
@@ -39,38 +46,59 @@ abstract class EvaBaseFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        baseDisposables += EventPipelines.changeFragmentBackgroundResource
+        val evaRadioBtn = (radioBtn as? EvaRadioBtn)
+
+        disposables += EventPipelines.changeFragmentBackgroundResource
                 .distinctUntilChanged()
                 .subscribe {
                     view.setBackgroundResource(it)
                 }
 
-        optionsBtn?.setOnClickListener {
-            EventPipelines.openOptionsDrawer.onNext(Unit)
+        if (evaRadioBtn != null) {
+            disposables += EventPipelines.playbackStartStopPause
+                    .observeOn(AndroidSchedulers.mainThread())
+                    .subscribe {
+                        when {
+                            it.playbackInfo?.isRadio == true
+                                    && it.player.playWhenReady
+                                    && it.player.playbackState == Player.STATE_READY -> {
+                                evaRadioBtn.radioBtnState = EvaRadioBtn.RadioBtnState.PLAYING
+                            }
+                            it.player.playWhenReady
+                                    && it.player.playbackState == Player.STATE_BUFFERING -> {
+                                if (evaRadioBtn.radioBtnState != EvaRadioBtn.RadioBtnState.NOT_PLAYING) {
+                                    evaRadioBtn.radioBtnState = EvaRadioBtn.RadioBtnState.BUFFERING
+                                }
+                            }
+                            else -> evaRadioBtn.radioBtnState = EvaRadioBtn.RadioBtnState.NOT_PLAYING
+                        }
+                    }
         }
 
-        radioBtn?.setOnClickListener {
-            EventPipelines.openRadio.onNext(Unit)
+        optionsBtn?.setOnClickListener {
+            EventPipelines.toggleOptionsDrawer.onNext(Unit)
         }
+
+        evaRadioBtn?.initialize()
     }
 
     override fun onDestroyView() {
+        disposables.clear()
         super.onDestroyView()
-        baseDisposables.clear()
     }
 
     override fun onDestroy() {
+        disposables.dispose()
         super.onDestroy()
-        baseDisposables.dispose()
     }
 
     protected fun safeReplaceDisposable(oldDisposable: Disposable?, newDisposable: Disposable?): Disposable? {
         if (oldDisposable != null) {
-            baseDisposables -= oldDisposable
-            oldDisposable.dispose() //do not rely on baseDisposables.remove()
+            disposables -= oldDisposable
+            oldDisposable.dispose() //do not rely on disposables.remove(oldDisposable)
         }
         if (newDisposable != null && !newDisposable.isDisposed) {
-            baseDisposables += newDisposable
+            disposables += newDisposable
         }
         return newDisposable
     }
