@@ -1,10 +1,8 @@
 package hr.bpervan.novaeva.fragments
 
-import android.content.Context
 import android.content.Intent
 import android.os.Bundle
-import android.os.Handler
-import android.support.design.widget.Snackbar
+import android.util.Log
 import android.view.ContextThemeWrapper
 import android.view.LayoutInflater
 import android.view.View
@@ -13,28 +11,27 @@ import androidx.core.net.toUri
 import androidx.core.os.bundleOf
 import androidx.core.view.isGone
 import androidx.core.view.isVisible
-import com.google.android.exoplayer2.extractor.DefaultExtractorsFactory
-import com.google.android.exoplayer2.source.ExtractorMediaSource
-import com.google.android.exoplayer2.upstream.DefaultBandwidthMeter
-import com.google.android.exoplayer2.upstream.DefaultDataSourceFactory
-import com.google.android.exoplayer2.util.Util
-import com.google.android.gms.analytics.HitBuilders
+import com.google.firebase.analytics.FirebaseAnalytics
 import hr.bpervan.novaeva.EventPipelines
 import hr.bpervan.novaeva.NovaEvaApp
 import hr.bpervan.novaeva.main.R
-import hr.bpervan.novaeva.model.EvaCategory
 import hr.bpervan.novaeva.model.EvaContent
 import hr.bpervan.novaeva.model.OpenContentEvent
-import hr.bpervan.novaeva.player.EvaPlayer
-import hr.bpervan.novaeva.services.novaEvaService
+import hr.bpervan.novaeva.rest.EvaDomain
+import hr.bpervan.novaeva.rest.NovaEvaService
+import hr.bpervan.novaeva.rest.serverByDomain
 import hr.bpervan.novaeva.storage.EvaContentDbAdapter
 import hr.bpervan.novaeva.storage.RealmConfigProvider
-import hr.bpervan.novaeva.util.EvaCache
-import hr.bpervan.novaeva.util.SCROLL_PERCENT_KEY
-import hr.bpervan.novaeva.util.networkRequest
+import hr.bpervan.novaeva.util.dataErrorSnackbar
+import hr.bpervan.novaeva.util.enumValueOrNull
+import hr.bpervan.novaeva.util.plusAssign
 import hr.bpervan.novaeva.util.sendEmailIntent
-import hr.bpervan.novaeva.views.*
-import io.reactivex.disposables.Disposable
+import hr.bpervan.novaeva.views.applyConfiguredFontSize
+import hr.bpervan.novaeva.views.applyEvaConfiguration
+import hr.bpervan.novaeva.views.loadHtmlText
+import io.reactivex.android.schedulers.AndroidSchedulers
+import io.reactivex.rxkotlin.subscribeBy
+import io.reactivex.schedulers.Schedulers
 import io.realm.Realm
 import kotlinx.android.synthetic.main.collapsing_content_header.view.*
 import kotlinx.android.synthetic.main.fragment_eva_content.*
@@ -46,16 +43,10 @@ class EvaContentFragment : EvaBaseFragment() {
 
     companion object : EvaFragmentFactory<EvaContentFragment, OpenContentEvent> {
 
-        private const val CONTENT_ID_KEY = "contentId"
-        private const val CATEGORY_ID_KEY = "categoryId"
-        private const val THEME_ID_KEY = "themeId"
-
         override fun newInstance(initializer: OpenContentEvent): EvaContentFragment {
             return EvaContentFragment().apply {
                 arguments = bundleOf(
-                        CONTENT_ID_KEY to initializer.contentMetadata.contentId,
-                        CATEGORY_ID_KEY to initializer.contentMetadata.categoryId,
-                        THEME_ID_KEY to initializer.themeId
+                        EvaFragmentFactory.INITIALIZER to initializer
                 )
             }
         }
@@ -63,133 +54,124 @@ class EvaContentFragment : EvaBaseFragment() {
 
     private lateinit var realm: Realm
 
-    public var contentId: Long = 0
-    private var categoryId: Long = 0
-    private var evaContent: EvaContent? = null
-
-    private val handler = Handler()
-
-    private var fetchFromServerDisposable: Disposable? = null
-        set(value) {
-            field = safeReplaceDisposable(field, value)
-        }
-
-    private var evaContentChangesDisposable: Disposable? = null
-        set(value) {
-            field = safeReplaceDisposable(field, value)
-        }
-
-    private var evaContentMetadataChangesDisposable: Disposable? = null
-        set(value) {
-            field = safeReplaceDisposable(field, value)
-        }
-
+    private lateinit var initializer: OpenContentEvent
+    private lateinit var domain: EvaDomain
     private var themeId = -1
+    public var contentId: Long = 0
+
+    private var evaContent: EvaContent? = null
+        set(value) {
+            field = value
+            if (value != null) {
+                updateUI(value)
+            }
+        }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
         val inState = savedInstanceState ?: arguments!!
-        contentId = inState.getLong(CONTENT_ID_KEY)
-        categoryId = inState.getLong(CATEGORY_ID_KEY)
-        themeId = inState.getInt(THEME_ID_KEY, -1)
 
-        savedInstanceState ?: NovaEvaApp.defaultTracker
-                .send(HitBuilders.EventBuilder()
-                        .setCategory("Vijesti")
-                        .setAction("OtvorenaVijest")
-                        .setLabel(contentId.toString())
-                        .setValue(contentId)
-                        .build())
+        initializer = inState.getParcelable(EvaFragmentFactory.INITIALIZER)!!
+
+        domain = initializer.domain
+        contentId = initializer.contentId
+        themeId = initializer.theme
 
         realm = Realm.getInstance(RealmConfigProvider.evaDBConfig)
-
-        if (savedInstanceState == null) {
-            fetchContentFromServer(contentId)
-        }
     }
 
     override fun onSaveInstanceState(outState: Bundle) {
-        outState.putLong(CONTENT_ID_KEY, contentId)
-        outState.putLong(CATEGORY_ID_KEY, categoryId)
-        outState.putInt(THEME_ID_KEY, themeId)
-        outState.putFloat(SCROLL_PERCENT_KEY, scrollView.calcScrollYPercent(scrollView.getChildAt(0).height))
-
+        outState.putParcelable(EvaFragmentFactory.INITIALIZER, initializer)
         super.onSaveInstanceState(outState)
     }
 
-    private fun createIfMissingAndSubscribeToEvaContentUpdates() {
-        EvaContentDbAdapter.createIfMissingEvaContentAsync(realm, contentId) {
-            subscribeToEvaContentUpdates()
-        }
-    }
+    private fun updateUI(evaContent: EvaContent) {
 
-    private fun subscribeToEvaContentUpdates() {
         view ?: return
 
-        evaContentChangesDisposable = EvaContentDbAdapter.subscribeToEvaContentUpdatesAsync(realm, contentId) { evaContent ->
+        evaCollapsingBar.collapsingToolbar.title = evaContent.title
 
-            evaCollapsingBar.collapsingToolbar.title = evaContent.contentMetadata!!.title
+        val coverImageInfo = evaContent.image
+        val coverImageView = evaCollapsingBar.coverImage
 
-            val coverImageInfo = evaContent.image
-            val coverImageView = evaCollapsingBar.coverImage
-
-            if (coverImageInfo != null) {
-                if (coverImageView != null) {
-                    imageLoader.displayImage(coverImageInfo.url, coverImageView)
-                }
-            } else {
-                //TODO bring back category headers
-//                val url = prefs.getString("hr.bpervan.novaeva.categoryheader." + contentData.contentMetadata!!.directoryId, null)
-//                if (url != null && headerImage != null) {
-//                    imageLoader.displayImage(url, headerImage)
-//                }
+        if (coverImageInfo != null) {
+            if (coverImageView != null) {
+                imageLoader.displayImage(coverImageInfo.url, coverImageView)
             }
-            vijestWebView.loadHtmlText(evaContent.text)
-
-            evaContent.audioURL?.let { audioUrl ->
-                imgMp3.setImageResource(R.drawable.vijest_ind_mp3_active)
-                if (audioUrl != this.evaContent?.audioURL) {
-                    prepareAudioStream(context!!, audioUrl, evaContent.contentId.toString(),
-                            evaContent.contentMetadata?.title ?: "nepoznato")
-                }
-                player_view?.apply {
-                    NovaEvaApp.evaPlayer.supplyPlayerToView(this, evaContent.contentId.toString())
-                    applyEvaConfiguration()
-                    requestFocus()
-                    showController()
-
-                }
+        } else {
+            val url = prefs.getString("hr.bpervan.novaeva.categoryheader." + evaContent.domain, null)
+            if (url != null && coverImageView != null) {
+                imageLoader.displayImage(url, coverImageView)
             }
-
-
-            evaContent.videoURL?.let { videoUrl ->
-                imgLink.setImageResource(R.drawable.vijest_ind_www_active)
-                imgLink.setOnClickListener {
-                    startActivity(Intent(Intent.ACTION_VIEW, videoUrl.toUri()))
-                }
-            }
-            if (evaContent.attachments.isNotEmpty()) {
-                imgText.setImageResource(R.drawable.vijest_ind_txt_active)
-                imgText.setOnClickListener {
-                    startActivity(Intent.createChooser(Intent(Intent.ACTION_VIEW, evaContent.attachments[0]!!.url.toUri()),
-                            "Otvaranje dokumenta " + evaContent.attachments[0]!!.name))
-                }
-            }
-
-            loadingCircle.isGone = true
-
-            this.evaContent = evaContent
         }
-    }
+        vijestWebView.loadHtmlText(evaContent.text)
 
-    private fun fetchContentFromServer(contentId: Long) {
-        fetchFromServerDisposable = novaEvaService.getContentData(contentId)
-                .networkRequest({ contentDataDTO ->
-                    EvaCache.cache(realm, contentDataDTO)
-                }, onError = {
-                    view?.snackbar(R.string.error_fetching_data, Snackbar.LENGTH_LONG)
-                })
+        evaContent.audioURL?.let { audioUrl ->
+            imgMp3.setImageResource(R.drawable.vijest_ind_mp3_active)
+
+            NovaEvaApp.evaPlayer.prepareAudioStream(
+                    audioUrl, evaContent.id.toString(),
+                    evaContent.title,
+                    isRadio = false,
+                    doAutoPlay = false,
+                    auth = serverByDomain(domain).auth)
+
+            player_view?.apply {
+                NovaEvaApp.evaPlayer.supplyPlayerToView(this, evaContent.id.toString())
+                applyEvaConfiguration()
+                requestFocus()
+                showController()
+
+            }
+            imgMp3.setOnClickListener {
+                startActivity(Intent(Intent.ACTION_VIEW, audioUrl.toUri()))
+            }
+        }
+
+
+        evaContent.videoURL?.let { videoUrl ->
+            imgLink.setImageResource(R.drawable.vijest_ind_www_active)
+            imgLink.setOnClickListener {
+                startActivity(Intent(Intent.ACTION_VIEW, videoUrl.toUri()))
+            }
+        }
+        if (evaContent.attachments.isNotEmpty()) {
+            imgText.setImageResource(R.drawable.vijest_ind_txt_active)
+            imgText.setOnClickListener {
+                startActivity(Intent.createChooser(Intent(Intent.ACTION_VIEW,
+                        evaContent.attachments[0]!!.url.toUri()), evaContent.attachments[0]!!.name))
+            }
+        }
+
+        loadingCircle.isGone = true
+
+        when (enumValueOrNull<EvaDomain>(evaContent.domain)) {
+            EvaDomain.VOCATION -> {
+                btnPoziv.isVisible = true
+                btnPoziv.setOnClickListener {
+                    sendEmailIntent(context,
+                            subject = getString(R.string.thinking_of_vocation),
+                            text = "Hvaljen Isus i Marija, javljam vam se jer razmišljam o duhovnom pozivu.",
+                            receiver = getString(R.string.vocation_email))
+                }
+            }
+            EvaDomain.ANSWERS -> {
+                btnPitanje.isVisible = true
+                btnPitanje.setOnClickListener {
+                    sendEmailIntent(context,
+                            subject = getString(R.string.having_a_question),
+                            text = getString(R.string.mail_preamble_praise_the_lord),
+                            receiver = getString(R.string.answers_email))
+                }
+            }
+            else -> {
+                /*nothing*/
+            }
+        }
+
+        vijestWebView.applyEvaConfiguration(prefs)
+        vijestWebView.loadHtmlText(evaContent.text)
     }
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
@@ -203,55 +185,51 @@ class EvaContentFragment : EvaBaseFragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
+        loadingCircle.isVisible = true
+
         EventPipelines.changeNavbarColor.onNext(R.color.Black)
         EventPipelines.changeStatusbarColor.onNext(R.color.VeryDarkGray)
         EventPipelines.changeFragmentBackgroundResource.onNext(R.color.White)
 
-        val savedScrollPercent = savedInstanceState?.getFloat(SCROLL_PERCENT_KEY, 0f) ?: 0f
-        if (savedScrollPercent > 0) {
-            vijestWebView.afterLoadAndLayoutComplete {
-                scrollView.scrollY = calcScrollYAbsolute(savedScrollPercent, scrollView.getChildAt(0).height)
-            }
+        disposables += EventPipelines.resizeText.subscribe {
+            vijestWebView?.applyConfiguredFontSize(prefs)
         }
 
-        initUI()
-    }
+        val evaContent = EvaContentDbAdapter.loadEvaContent(realm, contentId)
 
-    private fun initUI() {
-        val ctx = context ?: return
-
-        loadingCircle.isVisible = true
-
-        createIfMissingAndSubscribeToEvaContentUpdates()
-
-        vijestWebView.applyEvaConfiguration(prefs)
-        vijestWebView.loadHtmlText(evaContent?.text)
-
-        /** Is this 'Duhovni poziv' or 'Odgovori' category?  */
-        if (categoryId == EvaCategory.VOCATION.id.toLong()) {
-            btnPoziv.isVisible = true
-            btnPoziv.setOnClickListener {
-                val text = "Hvaljen Isus i Marija, javljam Vam se jer razmišljam o duhovnom pozivu."
-                sendEmailIntent(ctx, "Duhovni poziv", text, arrayOf("duhovnipoziv@gmail.com"))
-            }
+        if ((evaContent == null || savedInstanceState == null) && domain.isLegacy()) {
+            fetchContentFromServer_legacy(contentId)
+        } else {
+            this.evaContent = evaContent
         }
     }
 
-    private fun prepareAudioStream(context: Context, audioUri: String, contentId: String, contentTitle: String) {
-        val dataSourceFactory = DefaultDataSourceFactory(context,
-                Util.getUserAgent(context, resources.getString(R.string.app_name)),
-                DefaultBandwidthMeter())
+    override fun onResume() {
+        super.onResume()
 
-//        val factory = ExtractorMediaSource.Factory(dataSourceFactory).setCustomCacheKey(audioUri)
-//        val mediaSource = factory.createMediaSource(streamingUri)
-
-        NovaEvaApp.evaPlayer.prepareIfNeeded(EvaPlayer.PlaybackInfo(contentId, contentTitle), doAutoPlay = false) {
-            ExtractorMediaSource(audioUri.toUri(), dataSourceFactory, DefaultExtractorsFactory(), handler, null, audioUri)
-        }
+        FirebaseAnalytics.getInstance(requireContext())
+                .setCurrentScreen(requireActivity(), "Sadržaj '${initializer.title}'".take(36), "Content")
     }
 
     override fun onDestroy() {
         realm.close()
         super.onDestroy()
+    }
+
+    private fun fetchContentFromServer_legacy(contentId: Long) {
+        disposables += NovaEvaService.v2.getContentData(contentId)
+                .subscribeOn(Schedulers.io())
+                .observeOn(AndroidSchedulers.mainThread())
+                .subscribeBy(onSuccess = { contentDataDTO ->
+                    contentDataDTO.domain = domain
+                    EvaContentDbAdapter.addOrUpdateEvaContentAsync_legacy(realm, contentDataDTO) {
+                        evaContent = EvaContentDbAdapter.loadEvaContent(realm, contentId)
+                    }
+                }, onError = {
+                    Log.e("fetchContentLegacy", it.message, it)
+                    view?.dataErrorSnackbar()
+                    //load old
+                    evaContent = EvaContentDbAdapter.loadEvaContent(realm, contentId)
+                })
     }
 }
