@@ -1,35 +1,24 @@
 package hr.bpervan.novaeva.fragments
 
-import android.content.res.Configuration
 import android.os.Bundle
-import android.os.Handler
-import android.support.design.widget.Snackbar
-import android.support.v7.widget.DefaultItemAnimator
-import android.support.v7.widget.LinearLayoutManager
-import android.support.v7.widget.RecyclerView
-import android.util.Log
 import android.view.ContextThemeWrapper
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.core.content.edit
 import androidx.core.os.bundleOf
-import androidx.core.os.postDelayed
-import com.google.android.gms.analytics.HitBuilders
+import androidx.core.view.isVisible
+import com.google.firebase.analytics.FirebaseAnalytics
 import hr.bpervan.novaeva.EventPipelines
-import hr.bpervan.novaeva.util.EvaCache
 import hr.bpervan.novaeva.NovaEvaApp
 import hr.bpervan.novaeva.adapters.EvaRecyclerAdapter
 import hr.bpervan.novaeva.main.R
-import hr.bpervan.novaeva.model.*
-import hr.bpervan.novaeva.services.novaEvaService
+import hr.bpervan.novaeva.model.EvaDirectory
+import hr.bpervan.novaeva.model.OpenDirectoryEvent
+import hr.bpervan.novaeva.model.TIMESTAMP_FIELD
+import hr.bpervan.novaeva.rest.EvaDomain
 import hr.bpervan.novaeva.storage.EvaDirectoryDbAdapter
-import hr.bpervan.novaeva.storage.RealmConfigProvider
-import hr.bpervan.novaeva.util.NEW_CONTENT_KEY_PREFIX
-import hr.bpervan.novaeva.util.networkRequest
-import hr.bpervan.novaeva.views.snackbar
-import io.reactivex.disposables.Disposable
-import io.realm.Realm
+import hr.bpervan.novaeva.util.plusAssign
+import hr.bpervan.novaeva.util.sendEmailIntent
 import io.realm.Sort
 import kotlinx.android.synthetic.main.collapsing_directory_header.view.*
 import kotlinx.android.synthetic.main.fragment_directory_contents.*
@@ -38,125 +27,41 @@ import kotlinx.android.synthetic.main.top_izbornik.view.*
 /**
  * Created by vpriscan on 08.10.17..
  */
-class EvaDirectoryFragment : EvaBaseFragment() {
+class EvaDirectoryFragment : EvaAbstractDirectoryFragment() {
 
     companion object : EvaBaseFragment.EvaFragmentFactory<EvaDirectoryFragment, OpenDirectoryEvent> {
-
-        private const val directoryIdKey = "directoryId"
-        private const val directoryTitleKey = "directoryTitle"
-        private const val themeIdKey = "themeId"
 
         override fun newInstance(initializer: OpenDirectoryEvent): EvaDirectoryFragment {
             return EvaDirectoryFragment().apply {
                 arguments = bundleOf(
-                        directoryIdKey to initializer.directoryMetadata.directoryId,
-                        directoryTitleKey to initializer.directoryMetadata.title,
-                        themeIdKey to initializer.themeId
+                        EvaFragmentFactory.INITIALIZER to initializer
                 )
             }
         }
     }
 
+    private lateinit var initializer: OpenDirectoryEvent
 
-    private val handler = Handler()
-
-    private var fetchFromServerDisposable: Disposable? = null
-        set(value) {
-            field = safeReplaceDisposable(field, value)
-        }
-    private var evaDirectoryChangesDisposable: Disposable? = null
-        set(value) {
-            field = safeReplaceDisposable(field, value)
-        }
-
-    private var loadEvaDirectoryDisposable: Disposable? = null
-        set(value) {
-            field = safeReplaceDisposable(field, value)
-        }
-
-    private var directoryId: Long = -1
-    private lateinit var directoryTitle: String
     private var themeId: Int = -1
 
-    private lateinit var adapter: EvaRecyclerAdapter
-
-    private val elementsList: MutableList<TreeElementInfo> = mutableListOf()
-    private var hasMore = true
-    private var fetchingFromServer = true
-    private var loadingFromDb = true
-
-    private lateinit var realm: Realm
-
     override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
 
         val inState: Bundle = savedInstanceState ?: arguments!!
-        directoryId = inState.getLong(directoryIdKey)
-        directoryTitle = inState.getString(directoryTitleKey)
-        themeId = inState.getInt(themeIdKey)
+
+        initializer = inState.getParcelable(EvaFragmentFactory.INITIALIZER)!!
+
+        domain = initializer.domain
+        directoryId = initializer.directoryId
+        directoryTitle = initializer.title
+        themeId = initializer.theme
 
         adapter = EvaRecyclerAdapter(elementsList, { loadingFromDb || fetchingFromServer }, themeId)
-        adapter.registerAdapterDataObserver(DataChangeLogger())
 
-        savedInstanceState ?: NovaEvaApp.defaultTracker
-                .send(HitBuilders.EventBuilder()
-                        .setCategory("Direktorij")
-                        .setAction("OtvorenDirektorij")
-                        .setLabel(directoryTitle)
-                        .setValue(directoryId)
-                        .build())
-
-        realm = Realm.getInstance(RealmConfigProvider.evaDBConfig)
-
-        prefs.edit {
-            remove("$NEW_CONTENT_KEY_PREFIX$directoryId")
-        }
-
-        createIfMissingAndSubscribeToEvaDirectoryUpdates()
-
-        if (savedInstanceState == null) {
-            fetchEvaDirectoryDataFromServer()
-        }
-    }
-
-    private fun createIfMissingAndSubscribeToEvaDirectoryUpdates() {
-        EvaDirectoryDbAdapter.createIfMissingEvaDirectoryAsync(realm, directoryId, {
-            it.title = directoryTitle
-        }) {
-            subscribeToDirectoryUpdates()
-        }
-    }
-
-    private fun subscribeToDirectoryUpdates() {
-        evaDirectoryChangesDisposable = EvaDirectoryDbAdapter.subscribeToEvaDirectoryUpdatesAsync(
-                realm, directoryId, { evaDirectory ->
-
-            elementsList.clear()
-
-            val contentSorted = evaDirectory.contentMetadataList.sortedByDescending { it.timestamp }
-
-            val subdirectoriesSorted = evaDirectory.subDirectoryMetadataList//.sortedByDescending { todo ON SERVER }
-
-            if (contentSorted.size > 10) {
-                elementsList.addAll(contentSorted.take(10))
-                elementsList.addAll(subdirectoriesSorted)
-                elementsList.addAll(contentSorted.drop(10))
-            } else {
-                elementsList.addAll(contentSorted)
-                elementsList.addAll(subdirectoriesSorted)
-            }
-
-            loadingFromDb = false
-
-            adapter.notifyDataSetChanged()
-        })
+        super.onCreate(savedInstanceState)
     }
 
     override fun onSaveInstanceState(outState: Bundle) {
-        outState.putLong(directoryIdKey, directoryId)
-        outState.putString(directoryTitleKey, directoryTitle)
-        outState.putInt(themeIdKey, themeId)
-
+        outState.putParcelable(EvaFragmentFactory.INITIALIZER, initializer)
         super.onSaveInstanceState(outState)
     }
 
@@ -174,47 +79,72 @@ class EvaDirectoryFragment : EvaBaseFragment() {
         EventPipelines.changeStatusbarColor.onNext(R.color.VeryDarkGray)
         EventPipelines.changeFragmentBackgroundResource.onNext(R.color.White)
 
-        initUI()
-    }
-
-    private fun initUI() {
         evaDirectoryCollapsingBar.izbornikTop.izbornikTopNazivKategorije.apply {
             text = directoryTitle
             typeface = NovaEvaApp.openSansBold
         }
 
-        val recyclerView = evaRecyclerView as RecyclerView
+        val recyclerView = evaRecyclerView as androidx.recyclerview.widget.RecyclerView
 
-        val linearLayoutManager = LinearLayoutManager(context)
+        val linearLayoutManager = androidx.recyclerview.widget.LinearLayoutManager(context)
         recyclerView.layoutManager = linearLayoutManager
-        recyclerView.itemAnimator = DefaultItemAnimator()
+        recyclerView.itemAnimator = androidx.recyclerview.widget.DefaultItemAnimator()
         recyclerView.adapter = adapter
         recyclerView.addOnScrollListener(EndlessScrollListener(linearLayoutManager))
-    }
 
-    override fun onConfigurationChanged(newConfig: Configuration?) {
-        //A HACK TO DISPLAY CORRECT FRAGMENT VIEWS WHEN SWITCHING BETWEEN PORTRAIT AND LANDSCAPE
-        activity?.supportFragmentManager?.beginTransaction()?.detach(this)?.commitAllowingStateLoss()
-        super.onConfigurationChanged(newConfig)
-        activity?.supportFragmentManager?.beginTransaction()?.attach(this)?.commitAllowingStateLoss()
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-
-        realm.close()
-    }
-
-    class DataChangeLogger : RecyclerView.AdapterDataObserver() {
-        override fun onChanged() {
-            Log.d("recyclerDataChanged", "RecyclerView data changed")
+        when (domain) {
+            EvaDomain.VOCATION -> {
+                btnPoziv.isVisible = true
+                btnPoziv.setOnClickListener {
+                    sendEmailIntent(context,
+                            subject = getString(R.string.thinking_of_vocation),
+                            text = getString(R.string.mail_preamble_praise_the_lord)
+                                    + getString(R.string.mail_intro_vocation),
+                            receiver = getString(R.string.vocation_email))
+                }
+            }
+            EvaDomain.ANSWERS -> {
+                btnPitanje.isVisible = true
+                btnPitanje.setOnClickListener {
+                    sendEmailIntent(context,
+                            subject = getString(R.string.having_a_question),
+                            text = getString(R.string.mail_preamble_praise_the_lord),
+                            receiver = getString(R.string.answers_email))
+                }
+            }
+            else -> {/*nothing*/
+            }
         }
     }
 
-    inner class EndlessScrollListener(private val linearLayoutManager: LinearLayoutManager) : RecyclerView.OnScrollListener() {
+    override fun onResume() {
+        super.onResume()
+
+        FirebaseAnalytics.getInstance(requireContext())
+                .setCurrentScreen(requireActivity(), "Kategorija '$directoryTitle'".take(36), "Category")
+    }
+
+    override fun fillElements(evaDirectory: EvaDirectory) {
+        elementsList.clear()
+
+        val contentSorted = evaDirectory.contents.sortedByDescending { it.created }
+
+        val subdirectoriesSorted = evaDirectory.subCategories//.sortedByDescending { todo ON SERVER }
+
+        if (contentSorted.size > 10) {
+            elementsList.addAll(contentSorted.take(10))
+            elementsList.addAll(subdirectoriesSorted)
+            elementsList.addAll(contentSorted.drop(10))
+        } else {
+            elementsList.addAll(contentSorted)
+            elementsList.addAll(subdirectoriesSorted)
+        }
+    }
+
+    inner class EndlessScrollListener(private val linearLayoutManager: androidx.recyclerview.widget.LinearLayoutManager) : androidx.recyclerview.widget.RecyclerView.OnScrollListener() {
         private val visibleThreshold = 2
 
-        override fun onScrolled(recyclerView: RecyclerView, dx: Int, dy: Int) {
+        override fun onScrolled(recyclerView: androidx.recyclerview.widget.RecyclerView, dx: Int, dy: Int) {
             val visibleItemCount = recyclerView.childCount
             val totalItemCount = linearLayoutManager.itemCount
             val firstVisibleItem = linearLayoutManager.findFirstVisibleItemPosition()
@@ -222,62 +152,22 @@ class EvaDirectoryFragment : EvaBaseFragment() {
             if (firstVisibleItem > 0 && visibleItemCount > 0 && totalItemCount > 0) {
                 if (!fetchingFromServer && hasMore && totalItemCount - visibleItemCount <= firstVisibleItem + visibleThreshold) {
 
-                    /** Ako je zadnji u listi podkategorija, onda on nema UnixDatum, pa tražimo zadnji koji ima */
-
                     fetchingFromServer = true
                     refreshLoadingCircleState()
 
-                    loadEvaDirectoryDisposable = EvaDirectoryDbAdapter.loadEvaDirectoryAsync(realm, directoryId, { evaDirectory ->
-                        if (evaDirectory != null) {
-                            val oldestTimestamp = evaDirectory.contentMetadataList.sort(TIMESTAMP_FIELD, Sort.DESCENDING)
-                                    .lastOrNull()?.timestamp
-                            fetchEvaDirectoryDataFromServer(oldestTimestamp)
+                    if (domain.isLegacy()) {
+                        disposables += EvaDirectoryDbAdapter.loadEvaDirectoryAsync(realm, directoryId) { evaDirectory ->
+                            if (evaDirectory != null) {
+                                val oldestTimestamp = evaDirectory.contents.sort(TIMESTAMP_FIELD, Sort.DESCENDING)
+                                        .lastOrNull()?.created?.let { it / 1000 }
+                                fetchEvaDirectoryDataFromServer_legacy(oldestTimestamp)
+                            }
                         }
-                    })
+                    } else {
+                        fetchEvaDirectoryDataFromServer(pageOn + 1)
+                    }
                 }
             }
         }
-    }
-
-    private fun refreshLoadingCircleState() {
-        view?.let {
-            val loadingCircleIndex = adapter.itemCount - 1
-            it.post { adapter.notifyItemChanged(loadingCircleIndex) }
-        }
-    }
-
-    private fun fetchEvaDirectoryDataFromServer(timestamp: Long? = null) {
-        fetchingFromServer = true
-        refreshLoadingCircleState()
-
-        fetchFromServerDisposable = novaEvaService.getDirectoryContent(directoryId, timestamp)
-                .networkRequest({ evaDirectoryDTO ->
-                    loadingFromDb = true
-                    fetchingFromServer = false
-
-                    evaDirectoryDTO.directoryId = directoryId //todo fix on server
-
-                    EvaCache.cache(realm, evaDirectoryDTO)
-
-                    handler.postDelayed(4000) {
-                        /*if there are no actual new changes from server, data in cache will not be "updated"
-                        and on update callback (in subscribeToDirectoryUpdates) will not be called,
-                        resulting in progress circle spinning forever
-                        This block kills loading circle after some reasonable time
-                        */
-                        loadingFromDb = false
-                        refreshLoadingCircleState()
-                    }
-
-                    hasMore = evaDirectoryDTO.more > 0
-
-                }, onError = {
-                    handler.postDelayed(2000) {
-                        fetchingFromServer = false
-                        refreshLoadingCircleState()
-
-                        view?.snackbar(R.string.error_fetching_data, Snackbar.LENGTH_LONG)
-                    }
-                })
     }
 }
