@@ -5,15 +5,14 @@ import android.view.ContextThemeWrapper
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.google.android.gms.analytics.HitBuilders
+import com.google.firebase.analytics.FirebaseAnalytics
 import hr.bpervan.novaeva.EventPipelines
 import hr.bpervan.novaeva.NovaEvaApp
+import hr.bpervan.novaeva.main.BuildConfig
 import hr.bpervan.novaeva.main.R
-import hr.bpervan.novaeva.util.SCROLL_PERCENT_KEY
-import hr.bpervan.novaeva.views.afterLoadAndLayoutComplete
+import hr.bpervan.novaeva.util.plusAssign
+import hr.bpervan.novaeva.views.applyConfiguredFontSize
 import hr.bpervan.novaeva.views.applyEvaConfiguration
-import hr.bpervan.novaeva.views.calcScrollYAbsolute
-import hr.bpervan.novaeva.views.calcScrollYPercent
 import kotlinx.android.synthetic.main.collapsing_content_header.view.*
 import kotlinx.android.synthetic.main.fragment_simple_content.*
 
@@ -28,16 +27,6 @@ class EvaInfoFragment : EvaBaseFragment() {
         }
     }
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        savedInstanceState ?: NovaEvaApp.defaultTracker
-                .send(HitBuilders.EventBuilder()
-                        .setCategory("Informacije")
-                        .setAction("OtvoreneInformacije")
-                        .build())
-    }
-
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         return inflater.cloneInContext(ContextThemeWrapper(activity, R.style.AppTheme))
                 .inflate(R.layout.fragment_simple_content, container, false)
@@ -50,27 +39,24 @@ class EvaInfoFragment : EvaBaseFragment() {
         EventPipelines.changeStatusbarColor.onNext(R.color.VeryDarkGray)
         EventPipelines.changeFragmentBackgroundResource.onNext(R.color.White)
 
-        val savedScrollPercent = savedInstanceState?.getFloat(SCROLL_PERCENT_KEY, 0f) ?: 0f
-        if (savedScrollPercent > 0) {
-            webView.afterLoadAndLayoutComplete {
-                simpleContentScrollView.scrollY = calcScrollYAbsolute(savedScrollPercent, webView.height)
-            }
+        disposables += EventPipelines.resizeText.subscribe {
+            webView?.applyConfiguredFontSize(prefs)
         }
-
-        initUI()
-    }
-
-    override fun onSaveInstanceState(outState: Bundle) {
-        outState.putFloat(SCROLL_PERCENT_KEY, simpleContentScrollView.calcScrollYPercent(webView.height))
-        super.onSaveInstanceState(outState)
-    }
-
-    private fun initUI() {
 
         webView.applyEvaConfiguration(prefs)
 
         webView.loadUrl("file:///android_asset/info.html")
 
-        evaCollapsingBar.collapsingToolbar.title = "Nova Eva info"
+        val titleText = "${getString(R.string.app_name)} - v${BuildConfig.VERSION_NAME}" +
+                if (BuildConfig.DEBUG) " (debug)" else ""
+
+        evaCollapsingBar.collapsingToolbar.title = titleText
+    }
+
+    override fun onResume() {
+        super.onResume()
+
+        FirebaseAnalytics.getInstance(requireContext())
+                .setCurrentScreen(requireActivity(), "Informacije", "Information")
     }
 }
