@@ -1,5 +1,8 @@
 package hr.bpervan.novaeva.storage
 
+import hr.bpervan.novaeva.main.BuildConfig
+import hr.bpervan.novaeva.model.EvaDomainInfo
+import hr.bpervan.novaeva.rest.EvaDomain
 import io.realm.RealmConfiguration
 
 /**
@@ -10,8 +13,14 @@ object RealmConfigProvider {
 
     val evaDBConfig: RealmConfiguration by lazy {
         RealmConfiguration.Builder()
-                .name("novaEvaDb.realm")
-                .schemaVersion(1)
+                .name(BuildConfig.DB_NAME)
+                .schemaVersion(8)
+                .initialData { realm ->
+                    for (domain in EvaDomain.values()) {
+                        val domainInfo = EvaDomainInfo(domain.name, domain.domainEndpoint, domain.rootId)
+                        realm.insert(domainInfo)
+                    }
+                }
                 .deleteRealmIfMigrationNeeded()
                 .build()
     }
