@@ -2,48 +2,53 @@ package hr.bpervan.novaeva.util
 
 import android.content.Context
 import android.graphics.Bitmap
-
 import com.nostra13.universalimageloader.cache.disc.impl.UnlimitedDiskCache
 import com.nostra13.universalimageloader.cache.disc.naming.HashCodeFileNameGenerator
 import com.nostra13.universalimageloader.cache.memory.impl.LruMemoryCache
 import com.nostra13.universalimageloader.core.DisplayImageOptions
-import com.nostra13.universalimageloader.core.ImageLoader
 import com.nostra13.universalimageloader.core.ImageLoaderConfiguration
 import com.nostra13.universalimageloader.core.assist.ImageScaleType
 import com.nostra13.universalimageloader.core.decode.BaseImageDecoder
 import com.nostra13.universalimageloader.core.download.BaseImageDownloader
 import com.nostra13.universalimageloader.utils.StorageUtils
+import hr.bpervan.novaeva.rest.Server
+import java.net.HttpURLConnection
 
 object ImageLoaderConfigurator {
 
-    fun doInit(context: Context) {
+    fun createConfiguration(context: Context): ImageLoaderConfiguration {
         val cacheDir = StorageUtils.getCacheDirectory(context)
 
-        val config = ImageLoaderConfiguration.Builder(context)
-                .diskCacheExtraOptions(480, 800, null)
+        return ImageLoaderConfiguration.Builder(context)
+                .diskCacheExtraOptions(1080, 1080, null)
+                .diskCache(UnlimitedDiskCache(cacheDir))
+                .diskCacheFileCount(20)
+                .diskCacheFileNameGenerator(HashCodeFileNameGenerator())
+                .memoryCache(LruMemoryCache(4 * 1024 * 1024))
                 .denyCacheImageMultipleSizesInMemory()
-                .memoryCache(LruMemoryCache(2 * 1024 * 1024))
-                .diskCache(UnlimitedDiskCache(cacheDir)) // default
-                .diskCacheFileCount(100)
-                .diskCacheFileNameGenerator(HashCodeFileNameGenerator()) // default
-                .imageDownloader(BaseImageDownloader(context)) // default
-                .imageDecoder(BaseImageDecoder(true)) // default
-                .defaultDisplayImageOptions(createEvaDisplayImageOptions()) // default
+                .imageDecoder(BaseImageDecoder(true))
+                .imageDownloader(AuthImageDownloader(context))
+                .defaultDisplayImageOptions(DisplayImageOptions.Builder()
+                        .resetViewBeforeLoading(false)
+                        .cacheInMemory(true)
+                        .cacheOnDisk(false)
+                        .considerExifParams(false)
+                        .imageScaleType(ImageScaleType.IN_SAMPLE_POWER_OF_2)
+                        .bitmapConfig(Bitmap.Config.ARGB_8888)
+                        .extraForDownloader(Server.V3.auth)
+                        .build())
                 .writeDebugLogs()
                 .build()
-
-        ImageLoader.getInstance().init(config)
     }
 
-    private fun createEvaDisplayImageOptions(): DisplayImageOptions {
+    private class AuthImageDownloader(context: Context) : BaseImageDownloader(context) {
 
-        return DisplayImageOptions.Builder()
-                .resetViewBeforeLoading(false)
-                .cacheInMemory(true)
-                .cacheOnDisk(false)
-                .considerExifParams(false)
-                .imageScaleType(ImageScaleType.IN_SAMPLE_POWER_OF_2)
-                .bitmapConfig(Bitmap.Config.ARGB_8888)
-                .build()
+        override fun createConnection(url: String, extra: Any?): HttpURLConnection {
+            return super.createConnection(url, extra).apply {
+                if (extra is String) {
+                    setRequestProperty("Authorization", extra)
+                }
+            }
+        }
     }
 }
